pyproject.toml
[tool.poetry]
name = "PROJECT NAME"
version = "0.1.0"
description = "PROJECT DESCRIPTION"
authors = ["Name <contato@mail>"]

[tool.poetry.dependencies]
python = "^3.10.4"
tortoise-orm = {extras = ["asyncpg"], version = "^0.19.2"}
fastapi = "^0.79.0"
pydantic = {extras = ["dotenv", "email"], version = "^1.9.0"}

[tool.poetry.dev-dependencies]
black = "^22.6.0"
flake8 = "^4.0.1"
flake8-simplify = "^0.19.2"
flake8-bugbear = "^22.7.1"
wemake-python-styleguide = "^0.16.1"
flake8-docstrings = "^1.6.0"
pytest = "^7.1.2"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"

[tool.black]
line-length = 90
include = '\.pyi?$'
exclude = '''
/(
    \.git
| \.hg
| \.mypy_cache
| \.tox
| \.venv
| _build
| buck-out
| build
)/
'''

[tool.interrogate]
ignore-init-method = true
ignore-init-module = false
ignore-magic = false
ignore-semiprivate = false
ignore-private = false
ignore-property-decorators = false
ignore-module = true
ignore-nested-functions = false
ignore-nested-classes = true
ignore-setters = false
fail-under = 95
exclude = ["setup.py", "docs", "build"]
ignore-regex = ["^get$", "^mock_.*", ".*BaseClass.*"]
verbose = 0
quiet = false
whitelist-regex = []
color = true
generate-badge = "."
badge-format = "svg"

[tool.isort]
include_trailing_comma = true
use_parentheses = true
multi_line_output = 3
